requirements.txt
alembic==1.12.0
aniso8601==9.0.1
asgiref==3.7.2
attrs==23.2.0
blinker==1.6.2
boto3==1.35.0
botocore==1.35.0
certifi==2021.5.30
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
colormap==1.0.6
contourpy==1.0.7
coverage==4.4.1
cycler==0.11.0
distlib==0.3.7
Django==4.2.4
django-background-tasks==1.2.5
django-compat==1.0.15
django-paypal==2.0
django-serverless-cron==0.1.3
dnspython==2.4.2
easydev==0.13.1
et-xmlfile==1.1.0
filelock==3.12.2
Flask==2.3.3
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.39.4
gensim==4.3.2
greenlet==2.0.2
grpcio==1.65.5
grpcio-tools==1.65.5
numpy==1.24.3
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.1
paypal==1.2.5
pexpect==4.9.0
pika==1.3.2
